--- sleepingwar/src/main/java/com/joenastan/sleepingwar/plugin/Events/Tasks/DeleteWorldDelayed.java
@@ -1,29 +1,0 @@
-package com.joenastan.sleepingwar.plugin.Events.Tasks;
-
-import java.io.File;
-
-public class DeleteWorldDelayed implements Runnable {
-
-    private File worldFolder;
-
-    public DeleteWorldDelayed(File worldFolder) {
-        this.worldFolder = worldFolder;
-    }
-
-    @Override
-    public void run() {
-        deleteWorld(worldFolder);
-    }
-
-    private void deleteWorld(File f) {
-        if (f.isDirectory()) {
-            File[] files = f.listFiles();
-            for (File file : files) {
-                deleteWorld(file);
-            }
-        }
-
-        f.delete();
-    }
-
-}
