--- sleepingwar/src/main/java/com/joenastan/sleepingwar/plugin/Events/Tasks/PostEndedGame.java
@@ -1,18 +1,0 @@
-package com.joenastan.sleepingwar.plugin.Events.Tasks;
-
-import com.joenastan.sleepingwar.plugin.Game.SleepingRoom;
-
-public class PostEndedGame implements Runnable {
-
-    private SleepingRoom room;
-
-    public PostEndedGame(SleepingRoom room) {
-        this.room = room;
-    }
-
-    @Override
-    public void run() {
-        room.destroyRoom();
-    }
-
-}
