--- sleepingwar/src/main/java/com/joenastan/sleepingwar/plugin/Utility/VoidGenerator.java
@@ -1,22 +1,0 @@
-package com.joenastan.sleepingwar.plugin.Utility;
-
-import org.bukkit.Location;
-import org.bukkit.World;
-import org.bukkit.generator.ChunkGenerator;
-
-import java.util.Random;
-
-public class VoidGenerator extends ChunkGenerator {
-
-    @Override
-    public ChunkData generateChunkData(World world, Random random, int chunkX, int chunkZ, BiomeGrid biome) {
-        ChunkData chunk = createChunkData(world);
-        return chunk;
-    }
-
-    @Override
-    public Location getFixedSpawnLocation(World world, Random random) {
-        return new Location(world, 0, 10, 0);
-    }
-
-}
