+++ sleepingwar/src/main/java/com/joenastan/sleepingwar/plugin/commands/SleepingWarsPermissionsCommand.java
@@ -1,0 +1,15 @@
+package com.joenastan.sleepingwar.plugin.commands;
+
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandExecutor;
+import org.bukkit.command.CommandSender;
+import org.bukkit.event.Listener;
+
+public class SleepingWarsPermissionsCommand implements Listener, CommandExecutor {
+
+    @Override
+    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
+        return true;
+    }
+
+}
