+++ sleepingwar/src/main/java/com/joenastan/sleepingwar/plugin/events/Tasks/PostEndedGame.java
@@ -1,0 +1,18 @@
+package com.joenastan.sleepingwar.plugin.events.Tasks;
+
+import com.joenastan.sleepingwar.plugin.game.SleepingRoom;
+
+public class PostEndedGame implements Runnable {
+
+    private SleepingRoom room;
+
+    public PostEndedGame(SleepingRoom room) {
+        this.room = room;
+    }
+
+    @Override
+    public void run() {
+        room.destroyRoom();
+    }
+
+}
